// webapp/src/lib/api/model/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Kakeibo API
 * マルチテナント家計簿システム API
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryType } from "./categoryType";

export interface Category {
  id: string;
  workspaceId: string;
  /** @maxLength 100 */
  name: string;
  categoryType: CategoryType;
  /** @nullable */
  parentId?: string | null;
}