// webapp/src/lib/api/model/categoryType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Kakeibo API
 * マルチテナント家計簿システム API
 * OpenAPI spec version: 1.0.0
 */

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
} as const;